deleted sceneDelegate and main storyboard
