+++ moviesApp/moviesApp/Modules/Home/MainViewController.swift
@@ -1,0 +1,21 @@
+//
+//  MainViewController.swift
+//  moviesApp
+//
+//  Created by Edwin Weru on 30/05/2023.
+//
+
+import UIKit
+
+class MainViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        setUp()
+    }
+
+    private func setUp() {
+        title = "Main View"
+    }
+}
