+++ moviesApp/moviesApp/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
