--- moviesApp/moviesApp/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  moviesApp
-//
-//  Created by Edwin Weru on 30/05/2023.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
